lambda/requirements-test.txt
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
moto==5.0.0
respx==0.22.0
httpx==0.26.0
boto3==1.34.34
python-chess==1.999
